fix: restore product rendering and align docs with resolved remediation items
- fix backend product list failure by removing risky multi-collection entity graph usage
- add centralized exception logging (message + stacktrace) in global handler
- surface frontend product loading failures with explicit loading/error/empty states
- update API fetch flow to avoid silent fallback that hid failures

docs: remove completed findings from active reports
- security_vulnerability.md: keep only open risks, move implemented fixes to resolved section
- scalability_maintainability_report.md: remove completed scalability/API/model fixes, keep migration-strategy gap
- readability_recommendations.md: remove completed DTO/response/refactor items, keep remaining naming/docs tasks
- backend-remediation-plan.md: refresh to current status and note Flyway rollback

+++ backend-services/src/main/java/org/example/modules/auth/dto/UserDto.java
@@ -1,5 +1,6 @@
 package org.example.modules.auth.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -8,7 +9,8 @@ public record UserDto(
         String email,
         String username,
         List<String> roles,
-        Double userDiscountPercentage,
+        BigDecimal userDiscountPercentage,
         LocalDate userDiscountStartDate,
         LocalDate userDiscountEndDate
-) {}
+) {
+}
