+++ backend-services/src/main/java/org/example/modules/cart/dto/CartItemDiscountDto.java
@@ -1,11 +1,13 @@
 package org.example.modules.cart.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record CartItemDiscountDto(
         Long id,
         String description,
-        Double percentage,
+        BigDecimal percentage,
         LocalDate startDate,
         LocalDate endDate
-) {}
+) {
+}
