+++ backend-services/src/main/java/org/example/modules/cart/dto/CartItemDto.java
@@ -10,7 +10,8 @@ public record CartItemDto(
         String imageUrl,
         int quantity,
         CartItemDiscountDto productDiscount,
-        double userDiscountPercentage,
-        double employeeDiscountPercentage,
-        double totalDiscountPercentage
-) {}
+        BigDecimal userDiscountPercentage,
+        BigDecimal employeeDiscountPercentage,
+        BigDecimal totalDiscountPercentage
+) {
+}
